Use MUnit in AuthedContextSuite
